+++ test/solidlsp/r/test_r_basic.py
@@ -10,11 +10,10 @@ import pytest
 
 from solidlsp import SolidLanguageServer
 from solidlsp.ls_config import Language
-from test.conftest import is_ci
 from test.solidlsp.conftest import format_symbol_for_assert, has_malformed_name, request_all_symbols
 
 
-@pytest.mark.skipif(shutil.which("R") is None and not is_ci, reason="R is not available")
+@pytest.mark.skipif(shutil.which("R") is None, reason="R is not available")
 @pytest.mark.r
 class TestRLanguageServer:
     """Test basic functionality of the R language server."""